feat: myItem-service 수정

+++ src/main/java/com/plantify/item/service/MyItemServiceImpl.java
@@ -3,8 +3,10 @@ package com.plantify.item.service;
 import com.plantify.item.domain.dto.request.MyItemRequest;
 import com.plantify.item.domain.dto.response.MyItemResponse;
 import com.plantify.item.domain.entity.Item;
-import com.plantify.item.domain.entity.PurchaseItem;
+import com.plantify.item.domain.entity.MyItem;
 import com.plantify.item.global.exception.ApplicationException;
+import com.plantify.item.global.exception.errorcode.ItemErrorCode;
+import com.plantify.item.repository.ItemRepository;
 import com.plantify.item.repository.MyItemRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -22,8 +24,15 @@ public class MyItemServiceImpl implements MyItemService {
 
     @Override
     public List<MyItemResponse> getAllMyItems(String authorizationHeader) {
-        Long kakaoId = authenticationService.getUserInfo(authorizationHeader).kakaoId();
-        return myItemRepository.findByKakaoId(kakaoId)
+        if (authenticationService.validateAdminRole(authorizationHeader)) {
+            return myItemRepository.findAll()
+                    .stream()
+                    .map(MyItemResponse::from)
+                    .collect(Collectors.toList());
+        }
+
+        Long kakaoId = authenticationService.getKakaoId(authorizationHeader);
+        return myItemRepository.findByUserId(kakaoId)
                 .stream()
                 .map(MyItemResponse::from)
                 .collect(Collectors.toList());
@@ -31,37 +40,39 @@ public class MyItemServiceImpl implements MyItemService {
 
     @Override
     public MyItemResponse addMyItem(String authorizationHeader, MyItemRequest request) {
-        Long kakaoId = authenticationService.getUserInfo(authorizationHeader).kakaoId();
+        Long kakaoId = authenticationService.getKakaoId(authorizationHeader);
         Item item = itemRepository.findById(request.itemId())
                 .orElseThrow(() -> new ApplicationException(ItemErrorCode.ITEM_NOT_FOUND));
-        PurchaseItem myItem = request.toEntity(kakaoId, item);
-        PurchaseItem savedItem = myItemRepository.save(myItem);
+
+        MyItem myItem = request.toEntity(kakaoId, item);
+        MyItem savedItem = myItemRepository.save(myItem);
         return MyItemResponse.from(savedItem);
     }
 
     @Override
     public MyItemResponse updateMyItem(String authorizationHeader, Long myItemId, MyItemRequest request) {
-        Long kakaoId = authenticationService.getUserInfo(authorizationHeader).kakaoId();
-        PurchaseItem myItem = myItemRepository.findByKakaoIdAndMyItemId(kakaoId, myItemId)
+        MyItem myItem = myItemRepository.findById(myItemId)
                 .orElseThrow(() -> new ApplicationException(ItemErrorCode.ITEM_NOT_FOUND));
+        authenticationService.validateOwnership(authorizationHeader, myItem.getUserId());
 
-        PurchaseItem updatedItem = PurchaseItem.builder()
-                .purchaseItemId(myItem.getPurchaseItemId())
+        MyItem updatedItem = MyItem.builder()
+                .myItemId(myItem.getMyItemId())
                 .item(myItem.getItem())
                 .userId(myItem.getUserId())
-                .purchaseDate(myItem.getPurchaseDate())
                 .status(request.status())
+                .createdAt(myItem.getCreatedAt())
+                .updatedAt(myItem.getUpdatedAt())
                 .build();
 
-        PurchaseItem savedItem = myItemRepository.save(updatedItem);
+        MyItem savedItem = myItemRepository.save(updatedItem);
         return MyItemResponse.from(savedItem);
     }
 
     @Override
     public void deleteItem(String authorizationHeader, Long myItemId) {
-        Long kakaoId = authenticationService.getUserInfo(authorizationHeader).kakaoId();
-        PurchaseItem myItem = myItemRepository.findByKakaoIdAndMyItemId(kakaoId, myItemId)
+        MyItem myItem = myItemRepository.findById(myItemId)
                 .orElseThrow(() -> new ApplicationException(ItemErrorCode.ITEM_NOT_FOUND));
+        authenticationService.validateOwnership(authorizationHeader, myItem.getUserId());
         myItemRepository.delete(myItem);
     }
 }
